Add modernized ansible-pull example
A bit simpler, and self-contained solution with no extra files.

This is based on systemd timers and services, so it is has a smaller
footprint on systems which already have systemd installed (cron and
logrotate dependencies are gone). I assume that ansible is mostly used
with modern systems which have systemd available.

Systemd will restart the service every OnUnitInactiveSec=, while
making sure to not schedule overlapping runs. In practice, it is
convenient to change the timer to something as low as one minute when
quick updates are needed. With --only-if-changed the load on the
system is not even noticable because all that happens is a noop git
pull if no changes were made to the repo.

The logs end up in the journal, so no logging configuration is necessary.